Extract AbstractBytesReferenceTestCase
We have a ton of tests for PagedBytesReference but not really many for the other
implementation of BytesReference. This change factors out a basic AbstractBytesReferenceTestCase
that simplifies testing other implemenations. It also caught a coulpe of bug here and there like
a missing mask when reading bytes as ints in PagedBytesReference.
